feat: add perlin noise, improve physics and tick system

+++ src/input/Button.h
@@ -1,9 +1,11 @@
 #ifndef BUTTON_H
 #define BUTTON_H
 
+#include <cstdint>
+
 struct Button {
 	bool pressed = false;
-	float pressFrame = 0;
+	uint64_t pressFrame = 0;
 };
 
 #endif
